+++ internal/service/ssm/patch_baseline.go
@@ -7,6 +7,7 @@ import (
 	"regexp"
 	"strings"
 
+	"github.com/aws/aws-sdk-go-v2/service/ssm/types"
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/arn"
 	"github.com/aws/aws-sdk-go/service/ssm"
@@ -428,7 +429,7 @@ func resourcePatchBaselineDelete(ctx context.Context, d *schema.ResourceData, me
 	_, err := conn.DeletePatchBaseline(params)
 	if tfawserr.ErrCodeEquals(err, ssm.ErrCodeResourceInUseException) {
 		// Reset the default patch baseline before retrying
-		diags = append(diags, defaultPatchBaselineRestoreOSDefault(ctx, meta.(*conns.AWSClient), d.Get("operating_system").(string))...)
+		diags = append(diags, defaultPatchBaselineRestoreOSDefault(ctx, meta.(ssmClient), types.OperatingSystem(d.Get("operating_system").(string)))...)
 		if diags.HasError() {
 			return
 		}